+++ src/horus/util/model.py
@@ -136,7 +136,7 @@ class Mesh(object):
 
     def _calculate_normals(self):
         # Calculate the normals
-        tris = self.vertexes.reshape(self.vertex_count / 3, 3, 3)
+        tris = self.vertexes.reshape(int(self.vertex_count / 3), 3, 3)
         normals = np.cross(tris[::, 1] - tris[::, 0], tris[::, 2] - tris[::, 0])
         normals /= np.linalg.norm(normals)
         n = np.concatenate((np.concatenate((normals, normals), axis=1), normals), axis=1)
